Init SDL video appropriately for the platform
It seems that SDL is not completely consitent cross-platform in how the
SDL_INIT_VIDEO init works:
- on OSX it causes a crash if it is done at startup
- on linux it causes a crash if it is done as needed
- on windows it doesn't seem to care

It also appears that SDL FSO builds on windows were being given
resolutions enumerated by WIN32 not SDL so this behaviour was changed to
match the behaviour of JoyMan, by creating a new ResolutionMan and moving
a large chunk of code out of BasicSettingsPage.
