+++ code/tabs/BasicSettingsPage.cpp
@@ -41,6 +41,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #include "apis/SpeechManager.h"
 #include "apis/OpenALManager.h"
 #include "apis/JoystickManager.h"
+#include "apis/resolution_manager.hpp"
 #include "apis/HelpManager.h"
 #include "controls/ModList.h"
 #include "datastructures/FSOExecutable.h"
@@ -782,6 +783,7 @@ void BasicSettingsPage::OnFlagFileProcessingStatusChanged(wxCommandEvent& event)
 	if (status == FlagListManager::FLAG_FILE_PROCESSING_OK) {
 		this->SetupOpenALSection();
 		this->SetupJoystickSection();
+		this->SetUpResolution();
 	}
 }
 
@@ -1227,6 +1229,19 @@ void BasicSettingsPage::DisableExecutableChoiceControls(const ReasonForExecutabl
 }
 
 void BasicSettingsPage::SetUpResolution(const long minHorizRes, const long minVertRes) {
+	const ModItem* activeMod = ModList::GetActiveMod();
+	if (activeMod == NULL) {
+		// activeMod is only NULL when there is no FSO root available
+		// thus nothing to show on here either
+		// This would be better handled by not trying to set this combobox up when
+		// it can not possibly be needed.
+		return;
+	}
+	if (!FlagListManager::GetFlagListManager()->IsProcessingOK()) {
+		// We don't have a selected binary so we can't fill the resolution
+		// selection box correctly
+		return;
+	}
 	ProMan* proman = ProMan::GetProfileManager();
 	
 	wxChoice* resolutionCombo = dynamic_cast<wxChoice*>(
@@ -1236,15 +1251,7 @@ void BasicSettingsPage::SetUpResolution(const long minHorizRes, const long minVe
 	FillResolutionDropBox(resolutionCombo, minHorizRes, minVertRes);
 	
 	long width = 0, height = 0;
-	
-	const ModItem* activeMod = ModList::GetActiveMod();
-	if (activeMod == NULL) {
-		// activeMod is only NULL when there is no FSO root available
-		// thus nothing to show on here either
-		// This would be better handled by not trying to set this combobox up when
-		// it can not possibly be needed.
-		return;
-	}
+
 	const wxString& shortname(activeMod->shortname);
 	
 	bool hasValidRes = false;
@@ -1414,220 +1421,34 @@ void BasicSettingsPage::ShowSettings(const bool showSettings) {
 	}
 }
 
-class Resolution: public wxClientData {
-public:
-	Resolution(const int width, const int height, const bool isHeader) {
-		wxASSERT_MSG(width > 0, wxString::Format(_T("Resolution: provided width %d is invalid"), width));
-		wxASSERT_MSG(height > 0, wxString::Format(_T("Resolution: provided height %d is invalid"), height));
-		this->width = width;
-		this->height = height;
-		this->isHeader = isHeader;
-		resString = isHeader ? wxString::Format(_T("--- %d:%d ---"), width, height)
-			: wxString::Format(CFG_RES_FORMAT_STRING, width, height);
-	}
-	virtual ~Resolution() {}
-	const int GetWidth() const { return this->width; }
-	const int GetHeight() const { return this->height; }
-	const bool IsHeader() const { return this->isHeader; }
-	const wxString &GetResString() const { return this->resString; }
-	bool IsSameResolution(const int otherWidth, const int otherHeight) const {
-		return ((this->width == otherWidth) &&
-				(this->height == otherHeight));
-	}
-
-private:
-	int width;
-	int height;
-	bool isHeader;
-	wxString resString;
-};
-
-WX_DEFINE_ARRAY_PTR(Resolution *, ResolutionArray);
-
-// sort by aspect ratio, then by size, in descending order
-int CompareResolutions(Resolution **resp1, Resolution **resp2) {
-
-	wxASSERT_MSG(resp1 != NULL, _T("CompareResolutions: provided resp1 is NULL"));
-	wxASSERT_MSG((*resp1) != NULL, _T("CompareResolutions: resp1 points to NULL"));
-	wxASSERT_MSG(resp2 != NULL, _T("CompareResolutions: provided resp2 is NULL"));
-	wxASSERT_MSG((*resp2) != NULL, _T("CompareResolutions: resp2 points to NULL"));
-	
-	// compare the two width/height ratios by cross-multiplying and comparing the results
-	const int width1 = (*resp1)->GetWidth();
-	const int height1 = (*resp1)->GetHeight();
-	const int width2 = (*resp2)->GetWidth();
-	const int height2 = (*resp2)->GetHeight();
-	
-	wxASSERT_MSG(width1 > 0, wxString::Format(_T("CompareResolutions: width1 %d is invalid"), width1));
-	wxASSERT_MSG(height1 > 0, wxString::Format(_T("CompareResolutions: height1 %d is invalid"), height1));
-	wxASSERT_MSG(width2 > 0, wxString::Format(_T("CompareResolutions: width2 %d is invalid"), width2));
-	wxASSERT_MSG(height2 > 0, wxString::Format(_T("CompareResolutions: height2 %d is invalid"), height2));
-	
-	const int value1 = width1 * height2;
-	const int value2 = width2 * height1;
-	int result;
-	// first compare aspect ratio
-	if (value1 < value2) {
-		result = -1;
-	} else if (value1 > value2) {
-		result = 1;
-	} else { // then compare size if aspect ratios are equal
-		if (width1 < width2) {
-			result = -1;
-		} else if (width1 > width2) {
-			result = 1;
-		} else {
-			result = 0;
-		}
-	}
-
-	return (-1) * result; // for reverse comparison (descending order)
-}
-
-// brought to you by http://en.wikipedia.org/wiki/Euclidean_algorithm#Implementations
-int ComputeGCD(int a, int b) {
-	wxASSERT_MSG(a > 0, wxString::Format(_T("ComputeGCD(a=%d, b=%d): a must be positive"), a, b));
-	wxASSERT_MSG(b > 0, wxString::Format(_T("ComputeGCD(a=%d, b=%d): b must be positive"), a, b));
-
-//	part of the original algorithm, but irrelevant, since a and b are positive
-//	if (a == 0) {
-//		return b;
-//	}
-	while (b != 0) {
-		if (a > b) {
-			a = a - b;
-		} else {
-			b = b - a;
-		}
-	}
-	return a;
-}
-
-// Assumed resolutions has been sorted using CompareResolutions
-void AddHeaders(ResolutionArray &resolutions) {
-	if (resolutions.IsEmpty()) {
-		wxLogWarning(_T("AddHeaders: provided ResolutionArray is empty"));
-		return;
-	}
-
-	ResolutionArray headers;
-	wxArrayInt headerIndexes;
-
-	int lastAspectWidth = -1;
-	int lastAspectHeight = -1;
-	int width;
-	int height;
-	int gcd;
-	
-	// find aspect ratios and determine where the aspect ratio headers should be inserted
-	for (int i = 0, n = resolutions.GetCount(); i < n; ++i) {
-		width = resolutions.Item(i)->GetWidth();
-		height = resolutions.Item(i)->GetHeight();
-		gcd = ComputeGCD(width, height);
-		width /= gcd;
-		height /= gcd;
-		
-		// special exception: 8:5 should be 16:10
-		if ((width == 8) && (height == 5)) {
-			width *= 2;
-			height *= 2;
-		}
-		
-		if ((width != lastAspectWidth) || (height != lastAspectHeight)) {
-			wxLogDebug(_T(" found aspect ratio %d:%d"), width, height);
-			headers.Add(new Resolution(width, height, true));
-			headerIndexes.Add(i);
-			lastAspectWidth = width;
-			lastAspectHeight = height;
-		}
-	}
-	
-	// now insert the headers into resolutions
-	// must insert in reverse to avoid messing up insertion indexes
-	for (int i = headerIndexes.GetCount() - 1; i >= 0; --i) {
-		resolutions.Insert(headers.Item(i), headerIndexes.Item(i));
-	}
-}
-
 void BasicSettingsPage::FillResolutionDropBox(wxChoice *resChoice,
-	const long minHorizontalRes, const long minVerticalRes) {
-	
-	ResolutionArray resolutions;
-	
-	// first, detect supported resolutions
-#ifdef WIN32
-	DEVMODE deviceMode;
-	DWORD modeCounter = 0;
-	BOOL result;
+	const long minHorizontalRes, const long minVerticalRes)
+{
+	ResolutionMan::ResolutionArray resolutions;
+	ResolutionMan::ApiType apiType;
+	auto flagListManager = FlagListManager::GetFlagListManager();
 
-	wxLogDebug(_T("Enumerating graphics modes"));
-
-	do {
-		memset(&deviceMode, 0, sizeof(DEVMODE));
-		deviceMode.dmSize = sizeof(DEVMODE);
-
-		result = EnumDisplaySettings(NULL, modeCounter, &deviceMode);
-
-		if ( result == TRUE ) {
-			wxLogDebug(_T(" %dx%d %d bit %d hertz (%d)"),
-				deviceMode.dmPelsWidth,
-				deviceMode.dmPelsHeight,
-				deviceMode.dmBitsPerPel,
-				deviceMode.dmDisplayFrequency,
-				deviceMode.dmDisplayFlags);
-			// check to see if the resolution has already been added
-			// FIXME is there a reasonable way to not make populating resolutions O(n^2)?
-			bool resExists = false;
-			for (ResolutionArray::iterator it = resolutions.begin(), end = resolutions.end();
-				it != end; ++it) {
-				Resolution* p = (Resolution*)(*it);
-				if ((p)->IsSameResolution(deviceMode.dmPelsWidth, deviceMode.dmPelsHeight)) {
-					resExists = true;
-					break; // no need to keep looking
-				}
-			}
-
-			if ( !resExists &&
-					(deviceMode.dmPelsWidth >= static_cast<DWORD>(minHorizontalRes)) &&
-					(deviceMode.dmPelsHeight >= static_cast<DWORD>(minVerticalRes))) {
-				resolutions.Add(new Resolution(deviceMode.dmPelsWidth, deviceMode.dmPelsHeight, false));
-			}
+	if (flagListManager->IsProcessingOK()) {
+#if IS_WIN32
+		// If the current executable is a SDL exe, use that API
+		if (flagListManager->GetBuildCaps() & BUILD_CAP_SDL) {
+			apiType = ResolutionMan::API_SDL;
+		} else {
+			apiType = ResolutionMan::API_WIN32;
 		}
-		modeCounter++;
-	} while ( result == TRUE );
-#elif HAS_SDL == 1
-	wxLogDebug(_T("Enumerating graphics modes with SDL"));
-	// FSO currently only supports the primary display
-	const int DISPLAY_INDEX = 0;
-
-	if (SDL_GetNumVideoDisplays() > 0) {
-		int numDisplayModes = SDL_GetNumDisplayModes(DISPLAY_INDEX);
-
-		SDL_DisplayMode mode;
-
-		for (int i = 0; i < numDisplayModes; ++i) {
-			if (SDL_GetDisplayMode(DISPLAY_INDEX, i, &mode) != 0) {
-				wxLogWarning(_T("SDL_GetDisplayMode failed: %s"), SDL_GetError());
-				continue;
-			}
-
-			if ((mode.w >= minHorizontalRes) && (mode.h >= minVerticalRes)) {
-				resolutions.Add(new Resolution(mode.w, mode.h, false));
-			}
-		}
-	} else {
-		wxLogWarning(_T("SDL reported no displays!"));
-	}
 #else
-#error "BasicSettingsPage::FillResolutionDropBox not implemented because not on windows and SDL is not implemented"
+		// OSX and Linux always use the SDL api
+		apiType = ResolutionMan::API_SDL;
 #endif
-	// then arrange the resolutions for drop down box and insert them
-	resolutions.Sort(CompareResolutions);
-	AddHeaders(resolutions);
-	for (ResolutionArray::iterator it = resolutions.begin(), end = resolutions.end();
-		 it != end; ++it) {
-		Resolution* p = (Resolution*)(*it);
-		resChoice->Append(p->GetResString(), p);
+		ResolutionMan::EnumerateGraphicsModes(apiType, resolutions,
+			minHorizontalRes, minVerticalRes);
+
+		for (auto it = resolutions.begin(), end = resolutions.end();
+			it != end; ++it)
+		{
+			auto p = dynamic_cast<ResolutionMan::Resolution*>(*it);
+			resChoice->Append(p->GetResString(), p);
+		}
 	}
 }
 
@@ -1642,10 +1463,11 @@ int BasicSettingsPage::GetMaxSupportedResolution(const wxChoice& resChoice, long
 	
 	int maxResIndex = -1;
 	int maxResProduct = 0;
-	Resolution* res;
+	ResolutionMan::Resolution* res;
 
 	for (unsigned int i = 0; i < resChoice.GetCount(); ++i) {
-		res = dynamic_cast<Resolution*>(resChoice.GetClientObject(i));
+		res = dynamic_cast<ResolutionMan::Resolution*>(
+			resChoice.GetClientObject(i));
 		wxCHECK_MSG(res != NULL, wxNOT_FOUND,
 			wxString::Format(
 				_T("choice does not have Resolution object at index %u"), i));
@@ -1665,7 +1487,8 @@ int BasicSettingsPage::GetMaxSupportedResolution(const wxChoice& resChoice, long
 	wxCHECK_MSG(maxResIndex > -1, wxNOT_FOUND,
 		_T("maximum Resolution was not found"));
 	
-	res = dynamic_cast<Resolution*>(resChoice.GetClientObject(maxResIndex));
+	res = dynamic_cast<ResolutionMan::Resolution*>(
+		resChoice.GetClientObject(maxResIndex));
 	wxCHECK_MSG(res != NULL, wxNOT_FOUND,
 		_T("Choice is missing max Resolution object"));
 	width = res->GetWidth();
@@ -1684,7 +1507,7 @@ void BasicSettingsPage::OnSelectVideoResolution(
 	wxCHECK_RET( choice != NULL,
 		_T("Unable to find resolution combo"));
 
-	Resolution* res = dynamic_cast<Resolution*>(
+	auto res = dynamic_cast<ResolutionMan::Resolution*>(
 		choice->GetClientObject(choice->GetSelection()));
 	wxCHECK_RET( res != NULL,
 		_T("Choice does not have Resolution objects"));
@@ -1695,7 +1518,7 @@ void BasicSettingsPage::OnSelectVideoResolution(
 		choice->SetSelection(choice->GetSelection() + 1);
 	}
 	
-	res = dynamic_cast<Resolution*>(
+	res = dynamic_cast<ResolutionMan::Resolution*>(
 		choice->GetClientObject(choice->GetSelection()));
 	wxCHECK_RET( res != NULL,
 		_T("Cho2ce does not have Resolution objects"));
